Fixed up KitchenModelComparisonTest.py
It is now almost ready to run (the only thing left is to put it on CUDA). I've greatly simplified the process for adding more models. I've also told it to train on the whole dataset, instead of just brownies1. To run it, the only things you need to change are the numbers of epochs and layers and the hidden size. (You can find those by searching for "TODO" in the file)