Added a run on position data
Although it did better than previous runs, it was still 1-2 orders of magnitude worse than the benchmark. However, the benchmark is really good on positional data. This may be due to the fact that the framerate of the data is (supposedly) 120hz, and, therefore, assuming the subject never moves is a pretty good assumption.